Extract duplicate code from ContentSecurityPolicy and FeaturePolicy zendframework#176
